aamit/requirements.txt
Pillow
PyMuPDF
numpy
pandas
scikit-learn
tensorflow
geneticalgorithm
matplotlib
xgboost
seaborn
